+++ packages/vm/src/modules/file_system_cache.rs
@@ -1,10 +1,11 @@
 use std::fs;
+use std::hash::Hash;
 use std::io;
 use std::path::Path;
 use std::path::PathBuf;
 use thiserror::Error;
 
-use wasmer::{DeserializeError, Module, Store};
+use wasmer::{DeserializeError, Module, Store, Target};
 
 use crate::checksum::Checksum;
 use crate::errors::{VmError, VmResult};
@@ -40,14 +41,13 @@ use crate::modules::current_wasmer_module_version;
 ///   the module header version (<https://github.com/wasmerio/wasmer/issues/3193>). In cosmwasm-vm 1.1.0-1.1.1
 ///   the old value "v3" is still used along with Wasmer 2.3.0 (bug). From cosmwasm 1.1.2 onwards, this is
 ///   fixed by bumping to "v4".
-const MODULE_SERIALIZATION_VERSION: &str = "v4";
+/// - **v5**:<br>
+///   Version for cosmwasm_vm 1.3+ which adds a sub-folder with the target identier for the modules.
+const MODULE_SERIALIZATION_VERSION: &str = "v5";
 
 /// Representation of a directory that contains compiled Wasm artifacts.
 pub struct FileSystemCache {
-    /// The base path this cache operates in. Within this path, versioned directories are created.
-    /// A sophisticated version of this cache might be able to read multiple input versions in the future.
-    base_path: PathBuf,
-    wasmer_module_version: u32,
+    modules_path: PathBuf,
 }
 
 /// An error type that hides system specific error information
@@ -72,41 +72,37 @@ impl FileSystemCache {
     ///
     /// This method is unsafe because there's no way to ensure the artifacts
     /// stored in this cache haven't been corrupted or tampered with.
-    pub unsafe fn new(path: impl Into<PathBuf>) -> Result<Self, NewFileSystemCacheError> {
-        let wasmer_module_version = current_wasmer_module_version();
-
-        let path: PathBuf = path.into();
-        if path.exists() {
-            let metadata = path
+    pub unsafe fn new(base_path: impl Into<PathBuf>) -> Result<Self, NewFileSystemCacheError> {
+        let base_path: PathBuf = base_path.into();
+        if base_path.exists() {
+            let metadata = base_path
                 .metadata()
                 .map_err(|_e| NewFileSystemCacheError::CouldntGetMetadata)?;
-            if metadata.is_dir() {
-                if !metadata.permissions().readonly() {
-                    Ok(Self {
-                        base_path: path,
-                        wasmer_module_version,
-                    })
-                } else {
-                    Err(NewFileSystemCacheError::ReadonlyPath)
-                }
-            } else {
-                Err(NewFileSystemCacheError::ExistsButNoDirectory)
+            if !metadata.is_dir() {
+                return Err(NewFileSystemCacheError::ExistsButNoDirectory);
+            }
+            if metadata.permissions().readonly() {
+                return Err(NewFileSystemCacheError::ReadonlyPath);
             }
         } else {
             // Create the directory and any parent directories if they don't yet exist.
-            mkdir_p(&path).map_err(|_e| NewFileSystemCacheError::CouldntCreatePath)?;
-            Ok(Self {
-                base_path: path,
-                wasmer_module_version,
-            })
+            mkdir_p(&base_path).map_err(|_e| NewFileSystemCacheError::CouldntCreatePath)?;
         }
+
+        Ok(Self {
+            modules_path: modules_path(
+                &base_path,
+                current_wasmer_module_version(),
+                &Target::default(),
+            ),
+        })
     }
 
     /// Loads a serialized module from the file system and returns a module (i.e. artifact + store),
     /// along with the size of the serialized module.
     pub fn load(&self, checksum: &Checksum, store: &Store) -> VmResult<Option<(Module, usize)>> {
         let filename = checksum.to_hex();
-        let file_path = self.latest_modules_path().join(filename);
+        let file_path = self.modules_path.join(filename);
 
         let result = unsafe { Module::deserialize_from_file(store, &file_path) };
         match result {
@@ -130,12 +126,11 @@ impl FileSystemCache {
 
     /// Stores a serialized module to the file system. Returns the size of the serialized module.
     pub fn store(&mut self, checksum: &Checksum, module: &Module) -> VmResult<usize> {
-        let modules_dir = self.latest_modules_path();
-        mkdir_p(&modules_dir)
+        mkdir_p(&self.modules_path)
             .map_err(|_e| VmError::cache_err("Error creating modules directory"))?;
 
         let filename = checksum.to_hex();
-        let path = modules_dir.join(filename);
+        let path = self.modules_path.join(filename);
         module
             .serialize_to_file(&path)
             .map_err(|e| VmError::cache_err(format!("Error writing module to disk: {}", e)))?;
@@ -148,7 +143,7 @@ impl FileSystemCache {
     /// Returns true if the file existed and false if the file did not exist.
     pub fn remove(&mut self, checksum: &Checksum) -> VmResult<bool> {
         let filename = checksum.to_hex();
-        let file_path = self.latest_modules_path().join(filename);
+        let file_path = self.modules_path.join(filename);
 
         if file_path.exists() {
             fs::remove_file(file_path)
@@ -157,15 +152,6 @@ impl FileSystemCache {
         } else {
             Ok(false)
         }
-    }
-
-    /// The path to the latest version of the modules.
-    fn latest_modules_path(&self) -> PathBuf {
-        let version = format!(
-            "{}-wasmer{}",
-            MODULE_SERIALIZATION_VERSION, self.wasmer_module_version
-        );
-        self.base_path.join(version)
     }
 }
 
@@ -178,6 +164,27 @@ fn module_size(module_path: &Path) -> VmResult<usize> {
         .try_into()
         .expect("Could not convert file size to usize");
     Ok(module_size)
+}
+
+/// Creates an identifier for the Wasmer `Target` that is used for
+/// cache invalidation. The output is reasonable human friendly to be useable
+/// in file path component.
+fn target_id(target: &Target) -> String {
+    // Use a custom Hasher implementation to avoid randomization.
+    let mut deterministic_hasher = crc32fast::Hasher::new();
+    target.hash(&mut deterministic_hasher);
+    let hash = deterministic_hasher.finalize();
+    format!("{}-{:08X}", target.triple(), hash) // print 4 byte hash as 8 hex characters
+}
+
+/// The path to the latest version of the modules.
+fn modules_path(base_path: &Path, wasmer_module_version: u32, target: &Target) -> PathBuf {
+    let version_dir = format!(
+        "{}-wasmer{}",
+        MODULE_SERIALIZATION_VERSION, wasmer_module_version
+    );
+    let target_dir = target_id(target);
+    base_path.join(version_dir).join(target_dir)
 }
 
 #[cfg(test)]
@@ -252,11 +259,13 @@ mod tests {
         let module = compile(&wasm, None, &[]).unwrap();
         cache.store(&checksum, &module).unwrap();
 
-        let file_path = format!(
-            "{}/v4-wasmer1/{}",
+        let mut globber = glob::glob(&format!(
+            "{}/v5-wasmer1/**/{}",
             tmp_dir.path().to_string_lossy(),
             checksum
-        );
+        ))
+        .expect("Failed to read glob pattern");
+        let file_path = globber.next().unwrap().unwrap();
         let _serialized_module = fs::read(file_path).unwrap();
     }
 
@@ -288,5 +297,51 @@ mod tests {
         // Remove again
         let existed = cache.remove(&checksum).unwrap();
         assert!(!existed);
+    }
+
+    #[test]
+    fn target_id_works() {
+        let triple = wasmer::Triple {
+            architecture: wasmer::Architecture::X86_64,
+            vendor: target_lexicon::Vendor::Nintendo,
+            operating_system: target_lexicon::OperatingSystem::Fuchsia,
+            environment: target_lexicon::Environment::Gnu,
+            binary_format: target_lexicon::BinaryFormat::Coff,
+        };
+        let target = Target::new(triple.clone(), wasmer::CpuFeature::POPCNT.into());
+        let id = target_id(&target);
+        assert_eq!(id, "x86_64-nintendo-fuchsia-gnu-coff-4721E3F4");
+        // Changing CPU features changes the hash part
+        let target = Target::new(triple, wasmer::CpuFeature::AVX512DQ.into());
+        let id = target_id(&target);
+        assert_eq!(id, "x86_64-nintendo-fuchsia-gnu-coff-D5C8034F");
+
+        // Works for durrect target (hashing is deterministic);
+        let target = Target::default();
+        let id1 = target_id(&target);
+        let id2 = target_id(&target);
+        assert_eq!(id1, id2);
+    }
+
+    #[test]
+    fn modules_path_works() {
+        let base = PathBuf::from("modules");
+        let triple = wasmer::Triple {
+            architecture: wasmer::Architecture::X86_64,
+            vendor: target_lexicon::Vendor::Nintendo,
+            operating_system: target_lexicon::OperatingSystem::Fuchsia,
+            environment: target_lexicon::Environment::Gnu,
+            binary_format: target_lexicon::BinaryFormat::Coff,
+        };
+        let target = Target::new(triple, wasmer::CpuFeature::POPCNT.into());
+        let p = modules_path(&base, 17, &target);
+        assert_eq!(
+            p.as_os_str(),
+            if cfg!(windows) {
+                "modules\\v5-wasmer17\\x86_64-nintendo-fuchsia-gnu-coff-4721E3F4"
+            } else {
+                "modules/v5-wasmer17/x86_64-nintendo-fuchsia-gnu-coff-4721E3F4"
+            }
+        );
     }
 }